+++ src/Squirrel.CommandLine/OSX/Options.cs
@@ -2,9 +2,6 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.NET.HostModel.AppHost;
 
 namespace Squirrel.CommandLine.OSX
 {
